% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries}
\alias{countries}
\title{ISOs for world countries.}
\format{A data.table. Columns include:
\itemize{
  \item{name: country name. From United Nations sources (Terminology Bulletin Country
    Names and the Country and Region Codes for Statistical Use maintained by the United
    Nations Statistics Divisions).}
  \item{alpha 2: ISO 3166-1 two-letter alphabetic code.}
  \item{alpha 3: ISO 3166-1 three-letter alphabetic code.}
  \item{numeric: ISO 3166-1 three-digit numeric code.}
  \item{capital: country's capital city.}
}}
\source{
\url{http://www.iso.org/iso-3166-country-codes.html}.
}
\usage{
countries
}
\description{
ISO 3166-1 country codes.
}
\seealso{
\code{storethat} package from the \code{finRes} universe.
}
\keyword{datasets}
