% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{currencies}
\alias{currencies}
\title{ISOs for worldwide currencies.}
\format{A data.table. Columns include:
\itemize{
  \item{name: currency name.}
  \item{alphabetic: ISO 4217 three-letter alphabetic code.}
  \item{numeric: ISO 4217 three-digit numeric code.}
  \item{minor unit: number of digits after the decimal separator.}
  \item{country: country where the currency is issued. ISO 3166-1 - aplha 2 code.}
}}
\source{
\url{http://www.iso.org/iso-4217-currency-codes.html}.
}
\usage{
currencies
}
\description{
ISO 4217 codes for worldwide currencies. Both alphabetic and numeric codes are provided.
  The ISO 4217 three-letter alphabetic code standard is based on the ISO 3166 code standard for countries.
  The first two letters of the ISO 4217 alphabetic code are the same as the code for the country name,
  and where possible the third letter corresponds to the first letter of the currency name.
  The ISO 4217 three-digit numeric code is useful when currency codes need to be understood in countries
  that do not use Latin scripts and for computerised systems. Where possible the three-digit numeric code
  is the same as the numeric ISO 3166 country code.
}
\seealso{
\itemize{
  \item{\code{\link{countries}} dataset in this package.}
  \item{\code{storethat} package from the \code{finRes} universe.}
}
}
\keyword{datasets}
