% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exchanges}
\alias{exchanges}
\title{ISOs for worldwide financial exchanges.}
\format{
A data.table. Columns include:
\itemize{
  \item{name: institution name.}
  \item{MIC: ISO 10383 Market Identification Code.}
  \item{country: country where the exchange sits. ISO 3166-1 - aplha 2 code.}
  \item{city: city where the exchange sits.}
  \item{website: institution's website.}
}
}
\source{
\url{http://www.iso20022.org}.
}
\usage{
exchanges
}
\description{
The Market Identifier Code (MIC) is a unique identification code used to
  identify securities trading exchanges. I's a four alphanumeric character code, and
  is defined by the International Organization for Standardization (ISO) in ISO 10383
  which 'specifies a universal method of identifying exchanges, trading platforms and
  regulated or non-regulated markets as sources of prices and related information in
  order to facilitate automated processing'.
}
\seealso{
\itemize{
  \item{\code{\link{countries}} dataset in this package.}
  \item{\code{storethat} package from the \code{finRes} universe.}
}
}
\keyword{datasets}
