% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{currencies}
\alias{currencies}
\title{ISOs for worldwide currencies.}
\format{A data.table. Columns include:
\itemize{
  \item{name: currency name.}
  \item{alphabetic: ISO 4217 three-letter alphabetic code.}
  \item{numeric: ISO 4217 three-digit numeric code.}
  \item{minor unit: number of digits after the decimal separator.}
  \item{country: country where the currency is issued. ISO 3166-1 - aplha 2 code.}
}}
\source{
\url{http://www.iso.org/iso-4217-currency-codes.html}.
}
\usage{
currencies
}
\description{
The currencies dataset corresponds to the ISO 4217 standard published by the
International Organization for Standardization (ISO) that defines codes for worldwide
currencies and comes as a three-letter alphabetic as well as an alternative three-digit
numeric code, both provided in the dataset. The ISO 4217 three-letter alphabetic code
standard is based on the ISO 3166-1 code standard for countries with the first two letters
corresponding the ISO 3166-1 alpha-2 code for the country issuing the corresponding
currency and the third corresponding to the first letter of the currency name when possible.
Where possible the three-digit numeric code is the same as the ISO 3166-1 numeric code
for the issuing country.
}
\seealso{
\itemize{
  \item{\code{\link{countries}} dataset in this package.}
  \item{\code{storethat} package from the \code{finRes} universe.}
}
}
\keyword{datasets}
